Gather NWB fixtures
